+++ src/Plugin_Updater_Admin.php
@@ -20,11 +20,7 @@ if ( ! defined( 'ABSPATH' ) ) {
 class Plugin_Updater_Admin extends Settings {
 	use License_Actions;
 
-	/**
-	 * Variables.
-	 *
-	 * @var string|array
-	 */
+	// phpcs:disable Squiz.Commenting.VariableComment.Missing
 	protected $api_url     = null;
 	protected $api_data    = [];
 	protected $item_name   = null;
@@ -38,6 +34,7 @@ class Plugin_Updater_Admin extends Settings {
 	protected $cache_key   = null;
 	protected $strings     = null;
 	protected $data        = null;
+	// phpcs:enable
 
 	/**
 	 * Class constructor.
@@ -84,10 +81,7 @@ class Plugin_Updater_Admin extends Settings {
 
 		// Populate version fallback.
 		if ( empty( $config['version'] ) ) {
-			if ( ! function_exists( 'get_plugin_data' ) ) {
-				require_once ABSPATH . 'wp-admin/includes/plugin.php';
-			}
-			$plugin        = get_plugin_data( $config['file'] );
+			$plugin        = get_file_data( $config['file'], 'plugin' );
 			$this->version = $plugin['Version'];
 		}