+++ CHANGES.md
@@ -12,3 +12,4 @@
 * created Settings page for licenses instead of appending to menus
 * added ability to set updater only using `Init->updater( $config )`
 * update examples for `afragen/wp-dependency-installer:^3`
+* sanitize, escape & ignore